% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listgen.R
\name{listgen}
\alias{listgen}
\title{Generate a permuted list of students}
\usage{
listgen(
  students,
  filename = "listgen",
  folder = tempdir(),
  replications_in_rounds = 1,
  number_of_rounds = 19,
  cols_per_page = 2,
  rows_per_page = 31,
  max_lag = 5,
  max_iter = 10000,
  show_groups = TRUE,
  open = FALSE,
  open_with = "evince"
)
}
\arguments{
\item{students}{A data frame containing information about the students.
It must contain the following columns: \code{course}, \code{seminar}, \code{uco},
\code{last_name}, and \code{first_name}. Other columns are ignored.}

\item{filename}{The name of the output file without extension;
default is "listgen".}

\item{folder}{The folder where the output file will be saved.}

\item{replications_in_rounds}{How many times should each person be
present in one round.}

\item{number_of_rounds}{How many rounds.}

\item{cols_per_page}{Columns per page.}

\item{rows_per_page}{Rows per page.}

\item{max_lag}{How many ucos in a row cannot be the same.}

\item{max_iter}{How many iterations should be performed before
decreasing \code{max_lag} by 1.}

\item{show_groups}{A logical value indicating whether to show groups of
students. Default is \code{TRUE}.}

\item{open}{A logical value indicating whether to open the output file.}

\item{open_with}{The program to use to open the output file.}
}
\value{
None
}
\description{
\code{listgen()} generates a permuted list of students
for seminars based on the input data.
}
\examples{
\dontrun{
students <- tibble::tribble(
    ~course, ~seminar, ~uco, ~last_name, ~first_name,
    "BPE_AAA", "01", 123456, "Aloha", "Adam",
    "BPE_AAA", "01", 234567, "Brown", "Betty",
    "BPE_AAA", "01", 345678, "Cook", "Charlie",
    "BPE_AAA", "01", 456789, "Delacroix", "Diana"
)
listgen(students)
}
}
