% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listgen.R
\name{listgen}
\alias{listgen}
\title{Generate a permuted list of students}
\usage{
listgen(
  students,
  filename = "listgen",
  folder = tempdir(),
  replications_in_rounds = 1,
  number_of_rounds = 19,
  cols_per_page = 2,
  rows_per_page = 31,
  max_lag = 5,
  open = FALSE,
  open_with = "evince"
)
}
\arguments{
\item{students}{A data frame containing information about the students.}

\item{filename}{The name of the output file without extension;
default is "listgen".}

\item{folder}{The folder where the output file will be saved.}

\item{replications_in_rounds}{How many times should each person be
present in one round.}

\item{number_of_rounds}{How many rounds.}

\item{cols_per_page}{Columns per page.}

\item{rows_per_page}{Rows per page.}

\item{max_lag}{How many ucos in a row cannot be the same.}

\item{open}{A logical value indicating whether to open the output file.}

\item{open_with}{The program to use to open the output file.}
}
\value{
None
}
\description{
\code{listgen()} generates a permuted list of students
for seminars based on the input data.
}
\examples{
\dontrun{
students <- tibble::tribble(
    ~course, ~seminar, ~uco, ~last_name, ~first_name,
    "BPE_AAA", "01", 123456, "Novák", "Jan",
    "BPE_AAA", "01", 234567, "Novotný", "Petr",
    "BPE_AAA", "01", 345678, "Nováková", "Jana",
    "BPE_AAA", "01", 456789, "Novotná", "Petra"
)
listgen(students)
}
}
